+++ src/main/java/words/NegativeWord.java
@@ -1,20 +1,19 @@
-package model.words;
+package words;
 
 import java.util.HashSet;
 
-public class PositiveWord implements Word {
-    private static final int PRIORITY = 1;
+public class NegativeWord implements Word {
+    private static final int priority = 3;
     private final String word;
-    private HashSet<Integer> searchResultOfWord;
+    private  HashSet<Integer> searchResultOfWord;
 
-    public PositiveWord(String word) {
+    public NegativeWord(String word) {
         this.word = word.substring(1);
-
     }
 
     @Override
     public void filter(HashSet<Integer> userSearchResult) {
-        userSearchResult.retainAll(this.searchResultOfWord);
+        userSearchResult.removeAll(this.searchResultOfWord);
     }
 
     @Override
@@ -29,7 +28,6 @@ public class PositiveWord implements Word {
 
     @Override
     public int getPriority() {
-        return PRIORITY;
+        return priority;
     }
-
 }
